+++ CHANGELOG.md
@@ -8,13 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ##  version [1.0.0](https://github.com/i2bc/orfmine/releases) - 2022-11-25
 ### Added
-- output directory options for all orfmine callable scripts
 - softwares.ini file to inform on absolute local path to tango and/or iupred2a
 
 
 ### Changed
 - use of docker image through a handler process - commands to use docker image are the same as without docker (breaking)
+- output directory options for all orfmine callable scripts
 - rename output of orfplot (replace "output.png" by "outdir/filename.png")
+- place -o argument option in orfget with -outdir & -outname in orfget (not required by default)
 
 
 ##  version [0.8.7](https://github.com/i2bc/orfmine/releases) - 2022-11-12